pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "arman229", email = "armanashraf015@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.5.5",
    "fastapi[standard]>=0.115.12",
    "openai-agents>=0.0.11",
    "python-dotenv>=1.1.0",
    "uvicorn[standard]>=0.34.1",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[project.scripts]
chatbot = "chatbot:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
